Also accept JPEG images
Although they cannot have transparent parts, some teams will provide JPEG logos and it is a lot of work to convert them to PNGs with transparent background. Since the GameStateVisualizer only needs logos with transparent background if its own background is not white, JPEGs will be fine in most years.