Since 2016 picked up robots may return immediately in ready or set
I also did this for substituting robots since it makes no sense that
substituting robots is worse than picking one robot up and returning
another one.

However, if a robot is picked up when penalized, it still must not
return immediately. Otherwise, a robot that is penalized in Ready could
be picked up and returned immediately to avoid its penalty.